requirements.txt
requests 
xmltodict
pyyaml
python-dotenv